requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
drf-spectacular>=0.27
psycopg2-binary>=2.9
celery[redis]>=5.3
redis>=5.0
openpyxl>=3.1
gunicorn>=21.2
